+++ third-party/wocky-client/src/index.ts
@@ -5,6 +5,7 @@ export {Base, SERVICE_NAME} from './model/Base'
 export {Chat, IChat} from './model/Chat'
 export {Chats, IChats} from './model/Chats'
 export {Address} from './model/Address'
+export {Contact, IContact} from './model/Contact'
 export {Bot, IBot} from './model/Bot'
 export {BotPost, IBotPost} from './model/BotPost'
 export {Event, IEvent} from './model/Event'
@@ -12,7 +13,11 @@ export {EventBot, IEventBot} from './model/EventBot'
 export {EventBotCreate, IEventBotCreate} from './model/EventBotCreate'
 export {EventBotGeofence, IEventBotGeofence} from './model/EventBotGeofence'
 export {EventBotPost, IEventBotPost} from './model/EventBotPost'
-export {EventUserFollow, IEventUserFollow} from './model/EventUserFollow'
+export {
+  EventFriendInvite,
+  IEventFriendInvite,
+  IEventFriendInviteData,
+} from './model/EventFriendInvite'
 export {EventBotInvite, IEventBotInvite} from './model/EventBotInvite'
 export {File, IFile, FileRef} from './model/File'
 export {IFileService} from './transport/FileService'