+++ third-party/wocky-client/src/model/EventList.ts
@@ -3,7 +3,7 @@ import {EventBotCreate} from './EventBotCreate'
 import {EventBotPost} from './EventBotPost'
 import {EventBotGeofence} from './EventBotGeofence'
 import {EventDelete} from './EventDelete'
-import {EventUserFollow} from './EventUserFollow'
+import {EventFriendInvite} from './EventFriendInvite'
 import {EventBotInvite} from './EventBotInvite'
 import {createPaginable} from './PaginableList'
 import {IWocky} from '../index'
@@ -13,12 +13,12 @@ export const EventEntity = types.union(
   EventBotCreate,
   EventBotGeofence,
   EventDelete,
-  EventUserFollow,
+  EventFriendInvite,
   EventBotInvite
 )
 export type IEventEntity = typeof EventEntity.Type
 
-export function createEvent(params: any, service: IWocky) {
+export function createEvent(params: any, service: IWocky): IEventEntity {
   if (params.user) {
     params.user = service.profiles.get(params.user.id, params.user)
   }