+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Also: [Deployment history](https://github.com/hippware/tr-wiki/wiki/Client-deplo
 
 Ticket numbers refer to the ticket tracker for this project if not specified.
 
+# 4.0.2 - 2019 Jan 17
+
 * Comments are visible before the user followed the location (#3211)
 * [dev] wocky-client typing improvements, cleanup (PR #3215, PR #3217)
 * Fix chat message images (PR #3214)
@@ -18,6 +20,9 @@ Ticket numbers refer to the ticket tracker for this project if not specified.
 * Presence widget fixes (PR #3236)
   * New! Presence widget reappears after each kill/reload (#3219)
   * "Here" missing on widget (You're Here) (#2933)
+* [dev] upgrade CircleCI node version to 11 (#3244)
+* Bot creation missing autofocus to enter place or address (#3116)
+* New user relationship model (#3252)
 
 # 4.0.1 - 2018 Dec 20
 
@@ -43,6 +48,7 @@ Ticket numbers refer to the ticket tracker for this project if not specified.
   * [dev] Refactor Factory class (#3201)
   * wocky-client and messaging improvements (PR #3208, PR #3209)
   * Various improvements for images, bot posts (PR #3210)
+  * Fix missed avatar for new profiles (#3200)
 
 # 4.0.0 - 2018 Dec 10